+++ src/Data/PUS/CLTU.hs
@@ -1,3 +1,16 @@
+{-|
+Module      : Data.PUS.CLTU
+Description : Provides functions for CLTU (Command Link Transfer Unit) handling
+Copyright   : (c) Michael Oswald, 2019
+License     : BSD-3
+Maintainer  : michael.oswald@onikudaki.net
+Stability   : experimental
+Portability : POSIX
+
+This module is used for encoding and decoding to and from CLTU's. A TC Randomizer can be used, which is 
+basically an array of values which are xor'ed with the CLTU block data so that the distribution of 
+0's and 1's is more even (better for transmission to the satellite).
+-}
 {-# LANGUAGE BangPatterns
     , OverloadedStrings #-}
 module Data.PUS.CLTU
@@ -58,17 +71,17 @@ cltuTrailer n = B.replicate (fromIntegral n) 0x55
 
 {-# INLINABLE encode #-}
 -- | Encodes a CLTU into a ByteString suitable for sending via a transport protocol
--- | Takes a config as some values of the CLTU ecoding can be specified per mission
--- | (e.g. block length of the encoding)
+-- Takes a config as some values of the CLTU ecoding can be specified per mission
+-- (e.g. block length of the encoding)
 encode :: Config -> CLTU -> ByteString
 encode cfg cltu = encodeGeneric cfg cltu encodeCodeBlocks
 
 
 {-# INLINABLE encodeRandomized #-}
 -- | Encodes a CLTU into a ByteString suitable for sending via a transport protocol,
--- | but randomizes the data before
--- | Takes a config as some values of the CLTU ecoding can be specified per mission
--- | (e.g. block length of the encoding)
+-- but randomizes the data before
+-- Takes a config as some values of the CLTU ecoding can be specified per mission
+-- (e.g. block length of the encoding)
 encodeRandomized :: Config -> CLTU -> ByteString
 encodeRandomized cfg cltu = encodeGeneric cfg cltu encodeCodeBlocksRandomized
 
@@ -82,7 +95,7 @@ encodeGeneric cfg (CLTU pl) encoder = toLazyByteString (mconcat [lazyByteString 
 
 {-# INLINABLE decode #-}
 -- | Decodes incoming data into a CLTU. Returns either an error message or the decoded
--- | CLTU itself.
+-- CLTU itself.
 decode :: Config -> ByteString -> Either Text CLTU
 decode cfg pl = 
     if cltuHeader `B.isPrefixOf` pl
@@ -107,7 +120,7 @@ decode cfg pl =
 
 {-# INLINABLE decodeRandomized #-}
 -- | Decodes incoming data into a CLTU. Returns either an error message or the decoded
--- | CLTU itself.
+-- CLTU itself.
 decodeRandomized :: Config -> ByteString -> Either Text CLTU
 decodeRandomized cfg pl = 
     if cltuHeader `B.isPrefixOf` pl
@@ -133,9 +146,9 @@ decodeRandomized cfg pl =
 
 {-# INLINABLE encodeCodeBlocks #-}
 -- | Takes a ByteString as payload, splits it into CLTU code blocks according to 
--- | the configuration, calculates the parity for the code blocks by possibly 
--- | padding the last code block with the trailer and returns a builder
--- | with the result
+-- the configuration, calculates the parity for the code blocks by possibly 
+-- padding the last code block with the trailer and returns a builder
+-- with the result
 encodeCodeBlocks :: Config -> ByteString -> Builder
 encodeCodeBlocks cfg pl = 
     let cbSize = fromIntegral $ cfgCltuBlockSize cfg - 1
@@ -149,9 +162,9 @@ encodeCodeBlocks cfg pl =
 
 {-# INLINABLE encodeCodeBlocksRandomized #-}
 -- | Takes a ByteString as payload, splits it into CLTU code blocks according to 
--- | the configuration, applies randomization, calculates the parity for the code blocks by possibly 
--- | padding the last code block with the trailer and returns a builder
--- | with the result
+-- the configuration, applies randomization, calculates the parity for the code blocks by possibly 
+-- padding the last code block with the trailer and returns a builder
+-- with the result
 encodeCodeBlocksRandomized :: Config -> ByteString -> Builder
 encodeCodeBlocksRandomized cfg pl = 
     let cbSize = fromIntegral $ cfgCltuBlockSize cfg - 1
@@ -168,7 +181,7 @@ encodeCodeBlocksRandomized cfg pl =
 
 {-# INLINABLE encodeCodeBlock #-}
 -- | encodes a single CLTU code block. This function assumes that the given ByteString
--- | is already in the correct code block length - 1 (1 byte for parity will be added)
+-- is already in the correct code block length - 1 (1 byte for parity will be added)
 encodeCodeBlock :: ByteString -> Builder
 encodeCodeBlock block = 
     lazyByteString block <> word8 (cltuParity block)
@@ -176,7 +189,7 @@ encodeCodeBlock block =
 
 {-# INLINABLE encodeCodeBlockRandomized #-}
 -- | encodes a single CLTU code block and applies randomization. This function assumes that the given ByteString
--- | is already in the correct code block length - 1 (1 byte for parity will be added)
+-- is already in the correct code block length - 1 (1 byte for parity will be added)
 encodeCodeBlockRandomized :: Randomizer -> ByteString -> (Randomizer, Builder)
 encodeCodeBlockRandomized r block = 
     let (newR, rblock) = randomize r False block
@@ -187,7 +200,7 @@ encodeCodeBlockRandomized r block =
 
 {-# INLINABLE cltuParity #-}
 -- | calculates the parity of a single code block. The code block is assumed to
--- | be of the specified code block length
+-- be of the specified code block length
 cltuParity :: ByteString -> Word8 
 cltuParity !block =
     let proc :: Word8 -> Int32 -> Int32
@@ -201,8 +214,8 @@ cltuParity !block =
 
 {-# INLINABLE checkCodeBlock #-}
 -- | Checks a code block. First, it checks the length against the expected length,
--- | then checks the parity. Returns either an error message or the data block without
--- | the parity byte
+-- then checks the parity. Returns either an error message or the data block without
+-- the parity byte
 checkCodeBlock :: Word8 -> ByteString -> Either Text ByteString
 checkCodeBlock expectedLen block =
     let len = B.length block
@@ -230,8 +243,8 @@ checkCodeBlock expectedLen block =
 
 {-# INLINABLE checkCodeBlockRandomized #-}
 -- | Checks a code block. First, it checks the length against the expected length,
--- | then checks the parity. Returns either an error message or the data block without
--- | the parity byte
+-- then checks the parity. Returns either an error message or the data block without
+-- the parity byte
 checkCodeBlockRandomized :: Randomizer -> Word8 -> ByteString -> Either Text (Randomizer, ByteString)
 checkCodeBlockRandomized r expectedLen block =
     let len = B.length block