+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Revision history for esa-space-protocols
 
-## 0.1.0.0 -- YYYY-mm-dd
+## 0.1.0.0 -- 2019-05-01
 
 * First version. Released on an unsuspecting world.
+* Implemented CLTU encoding/decoding with TC Randomization
+* Added general functionality: chunking and hexdumps